wip, masterfilecheks
